Last emperor. Flame obsession
Heisei will have given the other choice Chosun to the fact that the prince does not have his own blood. In other words, there is no emperor's son in his own bloodline. At that time, a generation was born to the prince, and it became a man of the emperor's obsession.